move tag-multiselect to webpack bundle
