+++ templates/experiments/experiment_session_view.html
@@ -93,9 +93,10 @@
   {{ block.super  }}
   <script
     id="tag-multiselect"
-    src="{% static './tag_multiselect.js' %}"
+    src="{% static "js/tagMultiselect-bundle.js" %}"
     data-linkTagUrl="{% url 'annotations:link_tag' request.team.slug %}"
     data-unlinkTagUrl="{% url 'annotations:unlink_tag' request.team.slug %}"
   >
   </script>
+  <script>SiteJS.tagMultiselect.setupTagSelects();</script>
 {% endblock %}
